Added full docker support for run. Massive project refactor to make it easier in docker. Added .dockerignore as best practice.
