+++ src/dbt_project/models/staging/mtg_prices.sql
