+++ src/dbt_project/models/sub_sets/avg_price_of_cards.sql
