+++ src/dbt_project/models/sub_sets/daily_card_prices.sql
