+++ src/mtg_src/assets/3dgraph.py
