+++ src/resources/photos/example_dashboard.png
