+++ src/resources/photos/example_tree.png
